backend/requirements.txt
flask
pymongo
pandas
flask-cors
python-dotenv